- registered some of the API implementation as services - implemented create group and create project API calls

+++ src/main/kotlin/com/github/xeonkryptos/integration/gitlab/api/gitlab/model/GitlabUser.kt
@@ -13,7 +13,7 @@ import javax.imageio.ImageIO
  */
 @JsonIgnoreProperties(ignoreUnknown = true)
 data class GitlabUser @JsonCreator constructor(
-    @JsonProperty("id", required = true) val userId: Int,
+    @JsonProperty("id", required = true) val userId: Long,
     @JsonProperty("username", required = true) val username: String,
     @JsonProperty("name", required = true) val name: String,
     @JsonProperty("web_url", required = true) val server: String,
